Renamed file containing date range and date interval utils
