+++ src/settings/reducers/settings.ts
@@ -1,7 +1,7 @@
 import { createSlice, PayloadAction, PrepareAction } from '@reduxjs/toolkit';
 import { mergeDeepRight } from 'ramda';
 import { Theme } from '../../utils/theme';
-import { DateInterval } from '../../utils/dates/types';
+import { DateInterval } from '../../utils/helpers/dateIntervals';
 import { TagsOrder } from '../../tags/data/TagsListChildrenProps';
 import { ShortUrlsOrder } from '../../short-urls/data';
