+++ test/utils/dates/DateIntervalDropdownItems.test.tsx
@@ -1,6 +1,6 @@
 import { screen, waitFor } from '@testing-library/react';
 import { DateIntervalDropdownItems } from '../../../src/utils/dates/DateIntervalDropdownItems';
-import { DATE_INTERVALS, DateInterval, rangeOrIntervalToString } from '../../../src/utils/dates/types';
+import { DATE_INTERVALS, DateInterval, rangeOrIntervalToString } from '../../../src/utils/helpers/dateIntervals';
 import { DropdownBtn } from '../../../src/utils/DropdownBtn';
 import { renderWithEvents } from '../../__helpers__/setUpTest';
